+++ ontolutils/typing.py
@@ -98,3 +98,6 @@ TypeOrListOf: TypeAlias = Union[T, ResourceType, List[Union[T, ResourceType]]]
 OptionalTypeOrListOf: TypeAlias = Optional[Union[T, ResourceType, List[Union[T, ResourceType]]]]
 
 UnionResourceType: TypeAlias = Union[T, ResourceType]
+
+ResourceTypeOrListOf: TypeAlias = Union[ResourceType, List[ResourceType]]
+OptionalResourceTypeOrListOf: TypeAlias = Optional[Union[ResourceType, List[ResourceType]]]